requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.2
SQLAlchemy==2.0.36
stripe==10.3.0
python-multipart==0.0.9
python-slugify==8.0.4
